requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Caching==2.2.0
requests==2.32.3
python-dotenv==1.1.0
